feat: 🎸 Syntax Analysis
Adding quotes removal
