+++ include/minishell.h
@@ -70,6 +70,11 @@ int				find_dolar_pos(t_token *token_lst);
 int				is_dolar(char *value);
 int				var_mode(t_token **token_lst, char **value, char *value_start);
 int				is_word_join(char **abs_value, int is_start);
+int				var_mode(t_token **token_lst, char **value, char *value_start);
+int				quote_mode(t_token **token_lst, char **value,
+					char *quoted_value, char quote);
+int				regular_mode(t_token **token_lst, char **value, 
+					char *value_start, int i);
 void			token_clear(t_token *token);
 void			token_lst_clear(t_token **token_lst);
 void			token_lst_add_back(t_token **token_lst, t_token *token);
