+++ src/parser_token_build.c
@@ -68,34 +68,6 @@ int	append_token(t_token **token_lst, char **value,
 	return (1);
 }
 
-int	regular_mode(t_token **token_lst, char **value, char *value_start, int i)
-{
-	char	quote;
-
-	quote = 0;
-	while ((*value)[i] && !is_metacharacter((*value)[i]) && !is_dolar(*value))
-	{
-		if (!quote && ft_isspace((*value)[i]))
-			break ;
-		else if (quote && (*value)[i] == quote)
-		{
-			i++;
-			break ;
-		}
-		else if ((*value)[i] == '\'' || (*value)[i] == '\"')
-			quote = (*value)[i];
-		i++;
-		if (quote && !(*value)[i])
-		{
-			ft_putstr_fd("Unclosed quotes\n", 2);
-			return (0);
-		}
-	}
-	if (!append_token(token_lst, value, value_start, i))
-		return (0);
-	return (1);
-}
-
 int	default_build(t_token **token_lst,
 	char **value, char *value_start, int i)
 {
